Add Money class
- Change "unit" to "product_unit" in ProductTableView
- Switch all qreals to doubles
- Add currency class
- Add "Monies" classes, for grouping of money values
- Set app default locale to Nigeria (power to the people!)
- Add RecordTimestamp
- General refactoring
- Include 3rd party library for big decimal handling (Money class
dependency)

+++ src/rrcore/models/salepaymentmodel.cpp
@@ -29,7 +29,7 @@ QVariant SalePaymentModel::data(const QModelIndex &index, int role) const
 
     switch (role) {
     case AmountRole:
-        return m_salePayments.at(index.row()).amount;
+        return m_salePayments.at(index.row()).amount.toDouble();
     case MethodRole:
         return m_salePayments.at(index.row()).method.toString();
     case NoteRole:
