+++ src/rrcore/qmlapi/qmlpurchasereportmodel.cpp
@@ -26,7 +26,7 @@ QVariant QMLPurchaseReportModel::data(const QModelIndex &index, int role) const
     case UnitRole:
         return m_transactions.at(index.row()).product.unit.unit;
     case TotalAmountRole:
-        return m_transactions.at(index.row()).totalAmount;
+        return m_transactions.at(index.row()).totalAmount.toDouble();
     }
 
     return QVariant();
