+++ src/rrcore/utility/income/incometransaction.cpp
@@ -21,7 +21,7 @@ QVariantMap IncomeTransaction::toVariantMap() const
         { "expense_transaction_id", id },
         { "client_name", client.preferredName },
         { "purpose", purpose },
-        { "amount", amount }
+        { "amount", amount.toDouble() }
     };
 }
